// bookstore-webapp/src/generated/customers-service.ts
export interface components {
    schemas: {
        Customer: {
            id: string;
            name: string;
            email: string;
        };
        CustomerInput: {
            name: string;
            email: string;
        };
        Address: {
            id: string;
            line1: string;
            line2?: string;
            city: string;
            region: string;
            postalCode: string;
            country: string;
            isDefault?: boolean;
        };
        AddressInput: {
            line1: string;
            line2?: string;
            city: string;
            region: string;
            postalCode: string;
            country: string;
            isDefault?: boolean;
        };
        Error: {
            code: number;
            message: string;
            description?: string;
            moreInfo?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}