// bookstore-webapp/src/generated/notifications-service.ts
export interface components {
    schemas: {
        Notification: {
            id: string;
            orderId: string;
            message: string;
            read: boolean;
            createdAt: string;
        };
        Error: {
            code: number;
            message: string;
            description?: string;
            moreInfo?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}